Add real company house data csv, fix csv routes
- Add company_house_data.csv with first 10,000 rows of complete data set
- Fix LoadCSVData.scala
- Remove old company house models
- Move company model from SearchController to Models
