+++ app/controllers/v1/SearchController.scala
@@ -8,18 +8,9 @@ import javax.inject.Inject
 
 import com.typesafe.config.Config
 import play.api.Logger
-import models.{ CompanyHouseAddress, CompanyHouseJson, CompanyHouseObj }
+import models.{ Company, CompanyObj }
 import services.LoadCsvData
-import play.api.db._
-import org.apache.hadoop.fs.{ FileSystem, Path }
-import org.apache.spark.sql.SQLContext
-import java.sql.SQLException
-import java.sql.Connection
-import java.sql.ResultSet
-import java.sql.Statement
 import java.sql.DriverManager
-
-import play.api.libs.json.{ JsValue, Json, Writes };
 
 /**
  * Created by coolit on 18/07/2017.
@@ -44,9 +35,9 @@ class SearchController @Inject() (loadCsvData: LoadCsvData, val config: Config) 
     Action.async { implicit request =>
       Logger.info(s"Searching for company with id: ${companyNumber}")
       val res = companyNumber match {
-        case Some(companyNumber) if companyNumber.length > 0 => loadCsvData.ch.filter(_.company_number == companyNumber) match {
+        case Some(companyNumber) if companyNumber.length > 0 => loadCsvData.ch.filter(_.CompanyNumber == s""""$companyNumber"""") match {
           case Nil => NotFound(errAsJson(404, "not found", s"Could not find value ${companyNumber}")).future
-          case x => Ok(CompanyHouseObj.toJson(x(0))).future
+          case x => Ok(CompanyObj.toJson(x(0))).future
         }
         case _ => BadRequest(errAsJson(400, "missing parameter", "No query string found")).future
       }
@@ -117,102 +108,5 @@ class SearchController @Inject() (loadCsvData: LoadCsvData, val config: Config) 
         List()
       }
     }
-  }
-}
-
-case class Company(
-  CompanyName: String,
-  CompanyNumber: String,
-  CompanyCategory: String,
-  CompanyStatus: String,
-  CountryOfOrigin: String,
-  IncorporationDate: String,
-  // Address
-  AddressLine1: String,
-  AddressLine2: String,
-  PostTown: String,
-  County: String,
-  Postcode: String,
-  // Accounts
-  AccountRefDay: String,
-  AccountRefMonth: String,
-  AccountNextDueDate: String,
-  AccountLastMadeUpDate: String,
-  AccountCategory: String,
-  // Returns
-  ReturnsNextDueDate: String,
-  ReturnsLastMadeUpDate: String,
-  // SIC
-  SICCodeSicText1: String,
-  SICCodeSicText2: String,
-  SICCodeSicText3: String,
-  SICCodeSicText4: String
-)
-
-object CompanyObj {
-  implicit val writer = new Writes[Company] {
-    def writes(c: Company): JsValue = {
-      val sicText = List(c.SICCodeSicText1, c.SICCodeSicText2, c.SICCodeSicText3, c.SICCodeSicText4).filter(
-        _ != "\"\""
-      )
-      Json.obj(
-        "CompanyName" -> c.CompanyName,
-        "CompanyNumber" -> c.CompanyNumber,
-        "CompanyCategory" -> c.CompanyCategory,
-        "CompanyStatus" -> c.CompanyStatus,
-        "CountryOfOrigin" -> c.CountryOfOrigin,
-        "IncorporationDate" -> c.IncorporationDate,
-        "Address" -> Json.obj(
-          "AddressLine1" -> c.AddressLine1,
-          "AddressLine2" -> c.AddressLine2,
-          "PostTown" -> c.PostTown,
-          "County" -> c.County,
-          "Postcode" -> c.Postcode
-        ),
-        "Accounts" -> Json.obj(
-          "AccountRefDay" -> c.AccountRefDay,
-          "AccountRefMonth" -> c.AccountRefMonth,
-          "AccountNextDueDate" -> c.AccountNextDueDate,
-          "AccountLastMadeUpDate" -> c.AccountLastMadeUpDate,
-          "AccountCategory" -> c.AccountCategory
-        ),
-        "Returns" -> Json.obj(
-          "ReturnsNextDueDate" -> c.ReturnsNextDueDate,
-          "ReturnsLastMadeUpDate" -> c.ReturnsLastMadeUpDate
-        ),
-        "SICCodes" -> Json.obj(
-          "SicText" -> sicText
-        )
-      )
-    }
-  }
-
-  def toJson(company: Company): JsValue = {
-    Json.toJson(
-      Company(
-        company.CompanyName,
-        company.CompanyNumber,
-        company.CompanyCategory,
-        company.CompanyStatus,
-        company.CountryOfOrigin,
-        company.IncorporationDate,
-        company.AddressLine1,
-        company.AddressLine2,
-        company.PostTown,
-        company.County,
-        company.Postcode,
-        company.AccountRefDay,
-        company.AccountRefMonth,
-        company.AccountNextDueDate,
-        company.AccountLastMadeUpDate,
-        company.AccountCategory,
-        company.ReturnsNextDueDate,
-        company.ReturnsLastMadeUpDate,
-        company.SICCodeSicText1,
-        company.SICCodeSicText2,
-        company.SICCodeSicText3,
-        company.SICCodeSicText4
-      )
-    )
   }
 }
